requirements/dev.txt
-r base_requirements.txt
django-debug-toolbar==4.2.0
black==23.7.0
flake8==6.1.0
pre-commit==3.3.3
django-environ